CHANGE: move utility functions for mock and test functions to seperate file
